Relocate trajectory read/write to trajectory/io.py
